+++ server/advisors/advdata.cpp
@@ -1008,7 +1008,7 @@ void adv_best_government(struct player *pplayer)
         revolution_turns =
             get_player_bonus(pplayer, EFT_REVOLUTION_UNHAPPINESS);
         if (revolution_turns > 0) {
-          bonus -= 6 / revolution_turns;
+          bonus -= int(6 / revolution_turns);
         }
 
         val += (val * bonus) / 100;
